chore(cli): split analytics dispatch out of dispatch_surface_gap
dispatch_surface_gap.rs crossed the 500-line production module-size budget
when the graph-correlation epic added the topic_correlate command, turning
the CI 'Version Sync' job red on main. Move the analytics commands
(silent-hosts, clock-skew, anomalies, compare, apps) into a new
dispatch_surface_analytics.rs sibling. Both modules are now well under the
limit (gap 350, analytics 199). Pure code move, no behavior change.
