+++ server.json
@@ -7,11 +7,11 @@
     "url": "https://github.com/jmagar/cortex",
     "source": "github"
   },
-  "version": "1.31.3",
+  "version": "1.31.4",
   "packages": [
     {
       "registryType": "oci",
-      "identifier": "ghcr.io/jmagar/cortex:v1.31.3",
+      "identifier": "ghcr.io/jmagar/cortex:v1.31.4",
       "transport": {
         "type": "stdio"
       },
