+++ src/cli/dispatch_surface.rs
@@ -386,10 +386,12 @@ pub(crate) async fn run_notify_test(mode: &CliMode, args: NotifyTestArgs) -> Res
     Ok(())
 }
 
+pub(crate) use super::dispatch_surface_analytics::{
+    run_anomalies, run_apps, run_clock_skew, run_compare, run_silent_hosts,
+};
 pub(crate) use super::dispatch_surface_gap::{
-    run_anomalies, run_apps, run_clock_skew, run_compare, run_correlate_state, run_entity_lookup,
-    run_fleet_state, run_graph_around, run_graph_evidence, run_graph_explain, run_graph_rebuild,
-    run_graph_status, run_host_state, run_silent_hosts, run_topic_correlate,
+    run_correlate_state, run_entity_lookup, run_fleet_state, run_graph_around, run_graph_evidence,
+    run_graph_explain, run_graph_rebuild, run_graph_status, run_host_state, run_topic_correlate,
 };
 
 #[cfg(test)]
